requirements.txt
asgiref==3.10.0
Django==5.2.8
psycopg2-binary==2.9.11
pymodbus==3.11.3
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.16.1